fix: introduce lock around `deno info` calls to fix OOM (#50)
This fix builds on the insight in #55 but does so by performing the
minimal change required to fix the underlying issue. The reason for this
departure from #55 is due to our project's build failing when executed
using the proof of concept's plugin, presumably due to some unknown
change introduced on part of the proof of concept's extensive revision.

The fix reuses @notcome's lock component and threads said lock through
the codebase to ultimately place it around the `deno info --json ${id}`
calls.